dlcv-fall-2025-hw3-FrankLin-123712/requirements_p2.txt
imageio==2.36.1
matplotlib==3.9.2
numpy==1.26.4
Pillow==11.0.0
scipy==1.14.1
opencv-python==4.10.0.84
loralib==0.1.2
pycocotools==2.0.8
timm==1.0.9
torch==2.5.1
torchvision==0.20.1
pandas==2.2.3
tqdm==4.66.5
gdown==5.2.0
regex